requirements.txt
humanize
asyncpraw
dateparser
youtube_dl
python-dotenv
discord-together

git+https://github.com/Pycord-Development/pycord